Re=organize the code and rename classes
